app/pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = ""
authors = ["Roman <94143966+Carti23@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
loguru = "^0.7.2"
langchain = "^0.3.8"
uvicorn = "^0.32.1"
pytest = "^8.3.3"
redis = "^5.2.0"
openai = "^1.55.1"
langchain-openai = "^0.2.10"
pydantic-settings = "^2.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
